+++ src/App.vue
@@ -4,7 +4,6 @@ import { storyData } from '@/stores/story'
 
 const story = storyData()
 const installPrompt = ref(null)
-const storageStatus = ref(story.maxStorage < 1 ? true : false)
 
 window.addEventListener("beforeinstallprompt", e => {
   if (story.askInstall) {
@@ -27,71 +26,6 @@ const neverIntall = () => {
 }
 
 story.innerHeight = window.innerHeight
-
-/* Check The LocalStorage... */
-function getUsedSpaceOfLocalStorageInBytes() {
-    // Returns the total number of used space (in Bytes) of the Local Storage
-    var b = 0;
-    for (var key in window.localStorage) {
-        // eslint-disable-next-line no-prototype-builtins
-        if (window.localStorage.hasOwnProperty(key)) {
-            b += key.length + localStorage.getItem(key).length;
-        }
-    }
-    return b;
-}
-
-function getUnusedSpaceOfLocalStorageInBytes() {
-  var maxByteSize = 10485760; // 10MB
-  var minByteSize = 0;
-  var tryByteSize = 0;
-  var testQuotaKey = 'testQuota';
-  var timeout = 20000;
-  var startTime = new Date().getTime();
-  var unusedSpace = 0;
-  let runtime = 0
-  do {
-    runtime = new Date().getTime() - startTime;
-    try {
-      tryByteSize = Math.floor((maxByteSize + minByteSize) / 2);
-      //localStorage.setItem(testQuotaKey, new Array(tryByteSize).join('1'));
-      
-      //Recommended by @pkExec and @jrob007
-      localStorage.setItem(testQuotaKey, String('1').repeat(tryByteSize));
-      minByteSize = tryByteSize;
-    } catch (e) {
-      maxByteSize = tryByteSize - 1;
-    }
-  } while ((maxByteSize - minByteSize > 1) && runtime < timeout);
-
-  localStorage.removeItem(testQuotaKey);
-
-  if (runtime >= timeout) {
-    console.log("Unused Space Calculation May Be Off Due To Timeout.");
-  }
-
-  // Compensate for the byte size of the key that was used, then subtract 1 byte because the last value of the tryByteSize threw the exception
-  unusedSpace = tryByteSize + testQuotaKey.length - 1;
-  return unusedSpace;
-}
-
-function getLocalStorageQuotaInBytes() {
-    // Returns the total Bytes of Local Storage Space that the browser supports
-    var unused = getUnusedSpaceOfLocalStorageInBytes();
-    var used = getUsedSpaceOfLocalStorageInBytes();
-    var quota = unused + used;
-    story.currentStorage = used
-    story.maxStorage = quota
-    storageStatus.value = false
-    return quota;
-}
-
-if (storageStatus.value) {
-  getLocalStorageQuotaInBytes()
-} else {
-  story.currentStorage = getUsedSpaceOfLocalStorageInBytes()
-}
-/* End Check The LocalStorage... */
 </script>
 
 <template>